move assert cardinality checkpoint tests to kernel tests
